+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 maxscale cookbook CHANGELOG
 ---------------------------
 
+v1.2.0
+------
+- Florian Engelmann - add /etc/default/maxscale support to set ulimit nofile
+
+v1.1.0
+------
+- Florian Engelmann - add Maxscale 1.3.x support
+
 v1.0.0
 ------
 - Florian Engelmann - initial release